requirements.txt
black
gradio
pyyaml
requests
argparse